Changes to make upgrading work macbook easier.

+++ zshrc
@@ -26,15 +26,19 @@ getinteractive() {
 }
 
 upgradeoutdated() {
-  local sys=$(grep "^ID=" /etc/os-release | tr -d $'"') cmd="" bold=$(tput bold) reset=$(tput sgr0)
-  case "${sys##*=}" in
-    ("ubuntu") cmd="sudo DEBIAN_FRONTEND=noninteractive apt-get update --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get upgrade --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get autoclean --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get autoremove --yes" ;;
-    ("debian") cmd="sudo DEBIAN_FRONTEND=noninteractive apt-get update --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get upgrade --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get autoclean --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get autoremove --yes" ;;
-    ("amzn" | "fedora") cmd="sudo dnf update --assumeyes && sudo dnf clean all --assumeyes && sudo dnf autoremove --assumeyes" ;;
-    ("alpine") cmd="sudo apk update && sudo apk upgrade" ;;
-    ("freebsd") cmd="sudo pkg update && sudo pkg upgrade --yes && sudo freebsd-update fetch install" ;;
-    (*) cmd="echo 'Do not know how to upgrade this system'" ;;
-  esac
+  local os=$(grep "^ID=" /etc/os-release | tr -d $'"') cmd="" bold=$(tput bold) reset=$(tput sgr0) sys=$(uname -s | tr A-Z a-z)
+  if [[ "${sys}" == "darwin" ]]; then
+    cmd="sudo port -c -b -u upgrade outdated"
+  else
+    case "${os##*=}" in
+      ("ubuntu") cmd="sudo DEBIAN_FRONTEND=noninteractive apt-get update --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get upgrade --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get autoclean --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get autoremove --yes" ;;
+      ("debian") cmd="sudo DEBIAN_FRONTEND=noninteractive apt-get update --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get upgrade --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get autoclean --yes && sudo DEBIAN_FRONTEND=noninteractive apt-get autoremove --yes" ;;
+      ("amzn" | "fedora") cmd="sudo dnf update --assumeyes && sudo dnf clean all --assumeyes && sudo dnf autoremove --assumeyes" ;;
+      ("alpine") cmd="sudo apk update && sudo apk upgrade" ;;
+      ("freebsd") cmd="sudo pkg update && sudo pkg upgrade --yes && sudo freebsd-update fetch install" ;;
+      (*) cmd="echo 'Do not know how to upgrade this system'" ;;
+    esac
+  fi
   echo "Will run the command:"
   echo "${bold}${cmd//&& /&& \\ \n}${reset}\n\n"
   eval "${cmd}"
